# perf: optimize webstorage