--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -89,6 +89,7 @@ pin-project = "1.0.11" # don't pin because they yank crates from cargo
 rand = { version = "=0.8.5", features = ["small_rng"] }
 regex = "=1.6.0"
 ring = "=0.16.20"
+rusqlite = { version = "0.28.0", features = ["unlock_notify", "bundled"] }
 rustyline = { version = "=10.0.0", default-features = false, features = ["custom-bindings"] }
 rustyline-derive = "=0.7.0"
 secure_tempfile = { version = "=3.3.0", package = "tempfile" } # different name to discourage use in tests