--- a/ext/webstorage/Cargo.toml
+++ b/ext/webstorage/Cargo.toml
@@ -16,5 +16,5 @@ path = "lib.rs"
 [dependencies]
 deno_core = { version = "0.148.0", path = "../../core" }
 deno_web = { version = "0.97.0", path = "../web" }
-rusqlite = { version = "0.28.0", features = ["unlock_notify", "bundled"] }
+libsqlite3-sys = { version = "0.25.1", features = ["bundled"] }
 serde = { version = "1.0.136", features = ["derive"] }