`ext/webstorage` uses simple sqlite features. This commit removes rusqlite abstraction and directly interface with sqlite C API.

2.5x improvement in `localStorage.getItem()`

```
deno run -A --unstable cli/bench/webstorage.js 50 1000000 # main
time 2148 ms rate 465549
time 2120 ms rate 471698
time 2116 ms rate 472589
time 2112 ms rate 473484
time 2082 ms rate 480307
time 2079 ms rate 481000
```

```
target/release/deno run -A --unstable cli/bench/webstorage.js 50 1000000 # This patch
time 964 ms rate 1037344
time 966 ms rate 1035196
time 946 ms rate 1057082
time 942 ms rate 1061571
time 922 ms rate 1084598
time 931 ms rate 1074113
```